calloutDialog refactor - split code specific to image and link into their own super classes. Moved callout styles into a new stylesheet.
